+++ src/views/toutiaonew/components/copyTargeting.vue
@@ -55,152 +55,175 @@
             <div class="clear">
                 <div class="city_main">
                     <div class="city_title">定向名称</div>
+
                     <div class="city_box">
-                        <div class="item" v-for="item in TargetingList" @click="getTargeting(item)">{{item.targeting_name}}</div>
+                        <div class="item" v-for="item in TargetingList" @click="getTargeting(item)" :key="item.targeting_id">{{item.targeting_name}}</div>
                     </div>
+
                 </div>
                 <div class="city_main" style="margin-left:20px;">
                     <div class="city_title">{{targeting.targeting_name}}</div>
-                    <div class="city_box">
-                        <div class="padding-10 ">
-                            <div class="targeting_item" v-if="targeting.targeting.district">
-                                <span>地域：</span>
-                                <span class="grey">
-                                    <template v-for="item in toutiaoConfig.district" v-if="item.val_type==targeting.targeting.district">
-                                        {{item.name}}：
-                                    </template>
-                                    <template v-for="item in targeting.targeting.city">
-                                        <template v-for="subitem in province" v-if="item==subitem.value">
-                                            {{subitem.name}};
+                    <div style="position: relative;">
+                        <div class="city_box">
+                            <div class="padding-10 ">
+                                <div class="targeting_item" v-if="targeting.targeting.district">
+                                    <span>地域：</span>
+                                    <span class="grey">
+                                        <template v-for="item in toutiaoConfig.district" v-if="item.val_type==targeting.targeting.district">
+                                            {{item.name}}：
                                         </template>
-                                    </template>
-                                </span>
-                            </div>
-
-                            <div class="targeting_item" v-if="targeting.targeting.gender">
-                                <span>性别：</span>
-                                <span class="grey">
-                                    <template v-for="item in toutiaoConfig.gender" v-if="item.val_type==targeting.targeting.gender">
-                                        {{item.name}}
-                                    </template>
-                                </span>
-                            </div>
-
-                            <div class="targeting_item" v-if="targeting.targeting.age">
-                                <span>年龄：</span>
-                                <span class="grey">
-                                    <template v-for="item in toutiaoConfig.age">
-                                        <template v-for="subitem in targeting.targeting.age" v-if="item.val_type==subitem">
-                                            {{item.name}};
-                                        </template>
-                                    </template>
-                                </span>
-                            </div>
-
-                            <div class="targeting_item" v-if="targeting.targeting.ad_tag">
-                                <span>兴趣分类：</span>
-                                <span class="grey">
-                                    <template v-for="item in targeting.targeting.ad_tag">
-                                        <template v-for="subitem in ad_tag" v-if="item==subitem.value">
-                                            {{subitem.name}};
-                                            <template v-for="subTagsitem in subitem.subTags" v-if="item==subTagsitem.value">
-                                                {{subTagsitem.name}};
+                                        <template v-for="item in targeting.targeting.city">
+                                            <template v-for="subitem in province" v-if="item==subitem.value">
+                                                {{subitem.name}};
                                             </template>
                                         </template>
-                                    </template>
-                                </span>
-                            </div>
+                                    </span>
+                                </div>
 
-                            <div class="targeting_item" v-if="targeting.targeting.interest_tags">
-                                <span>兴趣关键词：</span>
-                                <span class="grey">
-                                    <template v-for="item in targeting.targeting.interest_tags">
-                                        {{item}};
-                                    </template>
-                                </span>
-                            </div>
-
-                            <div class="targeting_item" v-if="targeting.targeting.platform">
-                                <span>平台：</span>
-                                <span class="grey">
-                                    <template v-for="item in toutiaoConfig.platform">
-                                        <template v-for="subitem in targeting.targeting.platform" v-if="item.val_type==subitem">
-                                            {{item.name}};
+                                <div class="targeting_item" v-if="targeting.targeting.gender">
+                                    <span>性别：</span>
+                                    <span class="grey">
+                                        <template v-for="item in toutiaoConfig.gender" v-if="item.val_type==targeting.targeting.gender">
+                                            {{item.name}}
                                         </template>
-                                    </template>
-                                </span>
-                            </div>
+                                    </span>
+                                </div>
 
-                            <div class="targeting_item" v-if="targeting.targeting.ac">
-                                <span>网络类型：</span>
-                                <span class="grey">
-                                    <template v-for="item in toutiaoConfig.ac">
-                                        <template v-for="subitem in targeting.targeting.ac" v-if="item.val_type==subitem">
-                                            {{item.name}};
-                                        </template>
-                                    </template>
-                                </span>
-                            </div>
-
-                            <div class="targeting_item" v-if="targeting.targeting.carrier">
-                                <span>运营商：</span>
-                                <span class="grey">
-                                    <template v-for="item in toutiaoConfig.carrier">
-                                        <template v-for="subitem in targeting.targeting.carrier" v-if="item.val_type==subitem">
-                                            {{item.name}};
-                                        </template>
-                                    </template>
-                                </span>
-                            </div>
-
-                            <div class="targeting_item" v-if="targeting.targeting.activate_type">
-                                <span>新用户：</span>
-                                <span class="grey">
-                                    <template v-for="item in toutiaoConfig.activate_type">
-                                        <template v-for="subitem in targeting.targeting.activate_type" v-if="item.val_type==subitem">
-                                            {{item.name}};
-                                        </template>
-                                    </template>
-                                </span>
-                            </div>
-
-                            <div class="targeting_item" v-if="targeting.targeting.app_behavior_target!='NONE'&&targeting.targeting.app_category">
-                                <span>APP行为定向：</span>
-                                <span class="grey">
-                                    <template v-for="item in toutiaoConfig.app_behavior_target" v-if="item.val_type==targeting.targeting.app_behavior_target">
-                                        {{item.name}};
-                                    </template>
-                                    <template v-for="item in targeting.targeting.app_category">
-                                        <template v-for="subitem in app_category">
-                                            <template v-for="subTagsitem in subitem.subItems" v-if="item==subTagsitem.value">
-                                                {{subTagsitem.name}};
+                                <div class="targeting_item" v-if="targeting.targeting.age">
+                                    <span>年龄：</span>
+                                    <span class="grey">
+                                        <template v-for="item in toutiaoConfig.age">
+                                            <template v-for="subitem in targeting.targeting.age" v-if="item.val_type==subitem">
+                                                {{item.name}};
                                             </template>
                                         </template>
-                                    </template>
-                                </span>
-                            </div>
+                                    </span>
+                                </div>
 
-                            <div class="targeting_item" v-if="targeting.targeting.device_brand">
-                                <span>受众手机品牌：</span>
-                                <span class="grey">
-                                    <template v-for="item in targeting.targeting.device_brand">
-                                        <template v-for="subitem in device_brand" v-if="item==subitem.value">
-                                            {{subitem.name}};
+                                <div class="targeting_item" v-if="targeting.targeting.ad_tag">
+                                    <span>兴趣分类：</span>
+                                    <span class="grey">
+                                        <template v-for="item in targeting.targeting.ad_tag">
+                                            <template v-for="subitem in ad_tag" v-if="item==subitem.value">
+                                                {{subitem.name}};
+                                                <template v-for="subTagsitem in subitem.subTags" v-if="item==subTagsitem.value">
+                                                    {{subTagsitem.name}};
+                                                </template>
+                                            </template>
                                         </template>
-                                    </template>
-                                </span>
-                            </div>
+                                    </span>
+                                </div>
 
-                            <div class="targeting_item" v-if="targeting.targeting.article_category">
-                                <span>受众文章分类：</span>
-                                <span class="grey">
-                                    <template v-for="item in targeting.targeting.article_category">
-                                        <template v-for="subitem in article_category" v-if="item==subitem.value">
-                                            {{subitem.name}};
+                                <div class="targeting_item" v-if="targeting.targeting.interest_tags">
+                                    <span>兴趣关键词：</span>
+                                    <span class="grey">
+                                        <template v-for="item in targeting.targeting.interest_tags">
+                                            {{item}};
                                         </template>
-                                    </template>
-                                </span>
+                                    </span>
+                                </div>
+
+                                <div class="targeting_item" v-if="targeting.targeting.platform">
+                                    <span>平台：</span>
+                                    <span class="grey">
+                                        <template v-for="item in toutiaoConfig.platform">
+                                            <template v-for="subitem in targeting.targeting.platform" v-if="item.val_type==subitem">
+                                                {{item.name}};
+                                            </template>
+                                        </template>
+                                    </span>
+                                </div>
+
+                                <div class="targeting_item" v-if="targeting.targeting.ac">
+                                    <span>网络类型：</span>
+                                    <span class="grey">
+                                        <template v-for="item in toutiaoConfig.ac">
+                                            <template v-for="subitem in targeting.targeting.ac" v-if="item.val_type==subitem">
+                                                {{item.name}};
+                                            </template>
+                                        </template>
+                                    </span>
+                                </div>
+
+                                <div class="targeting_item" v-if="targeting.targeting.carrier">
+                                    <span>运营商：</span>
+                                    <span class="grey">
+                                        <template v-for="item in toutiaoConfig.carrier">
+                                            <template v-for="subitem in targeting.targeting.carrier" v-if="item.val_type==subitem">
+                                                {{item.name}};
+                                            </template>
+                                        </template>
+                                    </span>
+                                </div>
+
+                                <div class="targeting_item" v-if="targeting.targeting.activate_type">
+                                    <span>新用户：</span>
+                                    <span class="grey">
+                                        <template v-for="item in toutiaoConfig.activate_type">
+                                            <template v-for="subitem in targeting.targeting.activate_type" v-if="item.val_type==subitem">
+                                                {{item.name}};
+                                            </template>
+                                        </template>
+                                    </span>
+                                </div>
+
+                                <div class="targeting_item" v-if="targeting.targeting.app_behavior_target!='NONE'&&targeting.targeting.app_category">
+                                    <span>APP行为定向：</span>
+                                    <span class="grey">
+                                        <template v-for="item in toutiaoConfig.app_behavior_target" v-if="item.val_type==targeting.targeting.app_behavior_target">
+                                            {{item.name}};
+                                        </template>
+                                        <template v-for="item in targeting.targeting.app_category">
+                                            <template v-for="subitem in app_category">
+                                                <template v-for="subTagsitem in subitem.subItems" v-if="item==subTagsitem.value">
+                                                    {{subTagsitem.name}};
+                                                </template>
+                                            </template>
+                                        </template>
+                                    </span>
+                                </div>
+
+                                <div class="targeting_item" v-if="targeting.targeting.device_brand">
+                                    <span>受众手机品牌：</span>
+                                    <span class="grey">
+                                        <template v-for="item in targeting.targeting.device_brand">
+                                            <template v-for="subitem in device_brand" v-if="item==subitem.value">
+                                                {{subitem.name}};
+                                            </template>
+                                        </template>
+                                    </span>
+                                </div>
+
+                                <div class="targeting_item" v-if="targeting.targeting.article_category">
+                                    <span>受众文章分类：</span>
+                                    <span class="grey">
+                                        <template v-for="item in targeting.targeting.article_category">
+                                            <template v-for="subitem in article_category" v-if="item==subitem.value">
+                                                {{subitem.name}};
+                                            </template>
+                                        </template>
+                                    </span>
+                                </div>
+
+                                <div class="targeting_item" v-if="targeting.targeting.android_osv">
+                                    <span>受众最低android版本：</span>
+                                    <span class="grey">
+                                        <template v-for="item in toutiaoConfig.android_osv" v-if="item.val_type==targeting.targeting.android_osv">
+                                            {{item.name}}
+                                        </template>
+                                    </span>
+                                </div>
+
+                                <div class="targeting_item" v-if="targeting.targeting.ios_osv">
+                                    <span>受众最低ios版本：</span>
+                                    <span class="grey">
+                                        <template v-for="item in toutiaoConfig.ios_osv" v-if="item.val_type==targeting.targeting.ios_osv">
+                                            {{item.name}}
+                                        </template>
+                                    </span>
+                                </div>
                             </div>
+                            <Spin size="small" fix v-if="loading"></Spin>
                         </div>
                     </div>
                 </div>
@@ -220,6 +243,7 @@ export default {
             account_id: this.$route.query.account_id, //账户id
             toutiaoConfig: toutiaoConfig,
             byted_modal: false,
+            loading: false,
             TargetingList: "",
             targeting: {
                 account_id: "",
@@ -250,15 +274,18 @@ export default {
         },
         //获取定向详情
         getTargeting(item) {
+            this.loading = true;
             Axios.post('api.php', {
                 action: 'ttAdPut',
                 opt: 'getTargetingList',
                 targeting_id: item.targeting_id,
             }).then(res => {
+                this.loading = false;
                 if (res.ret == 1) {
                     this.targeting = res.data;
                 }
             }).catch(err => {
+                this.loading = false;
                 console.log('获取兴趣分类表失败' + err);
             })
         },
